Refactor for s3export

+++ fmod/base/util/array.py
@@ -69,4 +69,4 @@ def get_device():
 
 def array2tensor( darray: xa.DataArray ) -> Tensor:
     array_data: np.ndarray = np.ravel(darray.values).reshape( darray.shape )
-    return torch.tensor( array_data, device=get_device(), requires_grad=True )
+    return torch.tensor( array_data, device=get_device(), requires_grad=True, dtype=torch.float32 )
